+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - New flags for list: `--incs`, `--locals` and `--profiles`
+
+### Changed
+ - Rewrite man-page in reStructuredText
 
 ## [0.2.0] &ndash; 2021-02-07
 ### Added